updated with spdsw_r22142_2016-10-18 by cron

+++ idl/projects/maven/l2gen/muser_sta_l2gen.sh
@@ -24,8 +24,8 @@ setenv CDF_TMP /mydisks/home/maven
 setenv datestr `date +%Y%m%d%H%M%S`
 set suffix="$datestr"
 
-#check for lock file here
-if (! -e /tmp/STAL2lock.txt) then
+#check for lock file(s) here
+if (! -e /tmp/STAL2lock.txt && ! -e /tmp/STAL2Rlock.txt) then
     cd /mydisks/home/maven
     rm -f run_sta_l2gen.bm
     rm -f /mydisks/home/maven/stalog/run_sta_l2gen.txt$suffix
