+++ idl/projects/maven/l2gen/muser_sta_l2reprocess.sh
@@ -21,18 +21,14 @@ setenv IDL_PATH $IDL_PATH':'+$IDL_BASE_DIR
 # Set path for tmp files
 setenv CDF_TMP /mydisks/home/maven
 
-#check for lock file here
-if (! -e /tmp/STAL2Rlock.txt) then
+#check for lock file(s) here
+if (! -e /tmp/STAL2lock.txt && ! -e /tmp/STAL2Rlock.txt) then
     cd /mydisks/home/maven
     rm -f run_sta_l2reprocess.bm
     rm -f /tmp/run_sta_l2reprocess.txt
 
-    set line="rerun_sta_l2gen, ndays = 21"
+    set line="rerun_sta_l2gen"
     echo $line > run_sta_l2reprocess.bm
-    set line="rerun_sta_l2gen, ndays = 30"
-    echo $line >> run_sta_l2reprocess.bm
-    set line="rerun_sta_l2gen, ndays = 60"
-    echo $line >> run_sta_l2reprocess.bm
     echo exit >> run_sta_l2reprocess.bm
 
     idl run_sta_l2reprocess.bm > /tmp/run_sta_l2reprocess.txt &
